lisätty entity-testejä ja vaihdettu reference kansion nimi
