+++ tests/unit/reference_entity_test.py
@@ -2,7 +2,7 @@
 
 import unittest
 
-from viiteri.entities.reference import Article
+from viiteri.entities.references import Article, Book
 
 
 class TestReferenceEntity(unittest.TestCase):
@@ -12,6 +12,9 @@ class TestReferenceEntity(unittest.TestCase):
         self.test_article = Article(cite_key="petpet", author="Petteri",
                                     title="Petterin Kirja",
                                     journal="Petterin Kirjakokoelma", year="2003", volume="1")
+        self.test_book = Book(cite_key="petkir", author="Petteri", 
+                              editor="Petteri", title="Petterin kirja vol 2", 
+                              publisher="WSOY", year="2004")
 
     def test_article_constructor(self):
         """ Test that Article entity is correctly instantiated """
@@ -27,6 +30,19 @@ class TestReferenceEntity(unittest.TestCase):
 
         # Test that optional fields are set
         self.assertEqual(self.test_article.volume, "1")
+    
+    def test_book_constructor(self):
+        """ Test that Book-entity is correctly instantiated"""
+        # Test that abstract fields are set
+        self.assertEqual(self.test_book.type, "book")
+        self.assertEqual(self.test_book.cite_key, "petkir")
+
+        # Test that required fields are set
+        self.assertEqual(self.test_book.author, "Petteri")
+        self.assertEqual(self.test_book.editor, "Petteri")
+        self.assertEqual(self.test_book.title, "Petterin kirja vol 2")
+        self.assertEqual(self.test_book.publisher, "WSOY")
+        self.assertEqual(self.test_book.year, "2004")
 
     def test_init_article_with_missing_required_arguments(self):
         """ Test that Article entity raises ValueError with invalid arguments """
