--- viiteri/entities/reference/__init__.py
@@ -1,3 +1,0 @@
-""" viiteri/entities/reference/__init__.py """
-from viiteri.entities.reference.reference import Reference
-from viiteri.entities.reference.article import Article
