+++ viiteri/entities/references/__init__.py
@@ -1,0 +1,4 @@
+""" viiteri/entities/reference/__init__.py """
+from viiteri.entities.references.reference import Reference
+from viiteri.entities.references.article import Article
+from viiteri.entities.references.book import Book
