+++ viiteri/entities/references/article.py
@@ -1,29 +1,29 @@
-""" viiteri/entities/reference/book.py """
+""" viiteri/entities/reference/article.py """
 
 # pylint: disable=too-many-instance-attributes
-from viiteri.entities.reference import Reference
+from viiteri.entities.references import Reference
 
 
-class Book(Reference):
-    """ Class for representing 'book'-type references """
+class Article(Reference):
+    """ Class for representing 'article'-type references """
 
     def __init__(self, **kwargs):
-        if not kwargs.keys() >= {"cite_key", "author", "editor", "title", "publisher", "year"}:
+        if not kwargs.keys() >= {"cite_key", "author", "title", "journal", "year"}:
             raise ValueError("Missing required arguments")
 
         super().__init__("article", kwargs["cite_key"])
         self.author = kwargs["author"]
-        self.editor = kwargs["editor"]
         self.title = kwargs["title"]
-        self.publisher = kwargs["publisher"]
+        self.journal = kwargs["journal"]
         self.year = kwargs["year"]
 
         # Optional arguments
-        self.number = kwargs.get("number", None)
         self.volume = kwargs.get("volume", None)
+        self.number = kwargs.get("number", None)
         self.pages = kwargs.get("pages", None)
         self.month = kwargs.get("month", None)
-        self.note = kwargs.get("note", None)
         self.doi = kwargs.get("doi", None)
+        self.note = kwargs.get("note", None)
         self.issn = kwargs.get("issn", None)
-        self.isbn = kwargs.get("isbn", None)
+        self.zblnumber = kwargs.get("zblnumber", None)
+        self.eprint = kwargs.get("eprint", None)
