+++ viiteri/services/reference_service.py
@@ -1,7 +1,7 @@
 """ viiteri/services/reference_service """
 
 from viiteri.repositories.reference_repository import reference_repository as default_repository
-from viiteri.entities.reference import Reference
+from viiteri.entities.references import Reference
 from viiteri.utils.reference_factory import ReferenceFactory
